+++ src/helpers/main.js
@@ -11,3 +11,4 @@ export * from './table';
 export * from './history';
 export * from './array';
 export * from './object';
+export * from './string';